refactor(backend): split calculations, analytics, normalizer, report_generator under 500 LOC
- calculations.py 721 -> 491 + calculations_helpers.py (228) -- pure helpers
- analytics.py 687 -> 437 + analytics_helpers.py (262) -- time-range + SQL builders
- normalizer.py 617 -> 326 + normalizer_rows.py (335) -- row-level methods as mixin
- report_generator.py 524 -> 226 + report_generator_html.py (319) -- HTML rendering

Pure-helper extractions and mixin pattern; class APIs unchanged.
113 backend tests pass.

Frontend + backend now have ZERO files over 500 LOC.
